+++ src/main/java/net/sf/jabref/JabRef.java
@@ -45,7 +45,7 @@ import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.database.BibDatabaseMode;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.util.Util;
-import net.sf.jabref.wizard.auximport.AuxCommandLine;
+import net.sf.jabref.cli.AuxCommandLine;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
