+++ src/main/java/net/sf/jabref/gui/actions/NewSubDatabaseAction.java
@@ -8,7 +8,7 @@ import net.sf.jabref.gui.*;
 import net.sf.jabref.gui.util.PositionWindow;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.database.BibDatabaseMode;
-import net.sf.jabref.wizard.auximport.gui.FromAuxDialog;
+import net.sf.jabref.gui.auximport.FromAuxDialog;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
