+++ src/test/resources/net/sf/jabref/cli/origin.bib
