+++ src/test/resources/net/sf/jabref/cli/paper.aux
