+++ src/test/resources/net/sf/jabref/logic.auxparser/badpaper.aux
