+++ src/test/resources/net/sf/jabref/logic.auxparser/config.bib
