+++ src/test/resources/net/sf/jabref/logic.auxparser/nested.aux
