+++ src/test/resources/net/sf/jabref/logic.auxparser/origin.bib
@@ -1,0 +1,24 @@
+% Encoding: UTF-8
+
+@Book{Newton1999,
+  title =     {The Principia: mathematical principles of natural philosophy},
+  publisher = {Univ of California Press},
+  year =      {1999},
+  author =    {Newton, Isaac}
+}
+
+@Book{Darwin1888,
+  title =     {The descent of man, and selection in relation to sex},
+  publisher = {J. Murray},
+  year =      {1888},
+  author =    {Darwin, Charles}
+}
+
+@Book{Einstein1920,
+  title =     {Relativity: The special and general theory},
+  publisher = {Penguin},
+  year =      {1920},
+  author =    {Einstein, Albert}
+}
+
+@Comment{jabref-meta: DATABASE_TYPE:BibTeX;}
