+++ src/test/resources/net/sf/jabref/logic.auxparser/paper.aux
@@ -1,0 +1,8 @@
+\relax 
+\citation{Darwin1888}
+\citation{Einstein1920}
+\bibstyle{plain}
+\bibdata{origin}
+\bibcite{Darwin1888}{1}
+\bibcite{Einstein1920}{2}
+\@writefile{toc}{\contentsline {section}{\numberline {1}}{1}}
