+++ src/test/resources/net/sf/jabref/logic.auxparser/result.bib
